refactor(config): require explicit router group intent, fold into router_args
Three follow-ups to the shared router group.

`RouterArgGroup`'s arguments are now required keyword-only. They had
frontend-shaped defaults, so `RouterArgGroup()` -- the obvious call to copy
from the frontend -- gave a worker `--router-mode round-robin` and produced a
card. Since a card replaces the frontend's configuration wholesale, such a
worker would silently override a frontend running any other mode: no error,
just routing that ignores the operator. `tests/router/counter_worker.py`
already used that bare form, so the shape was reachable in one line. Requiring
both arguments turns forgetting into a startup TypeError, and a test pins it.

The frontend-only registrations now sit under a single `include_frontend_only`
guard instead of three interleaved ones, so what a worker does not get is
readable in one place.

`worker_router_args.py` is folded into `router_args.py`. Every other module
under `groups/` colocates its config class with its arg group, and the split
module held a config class and no group of its own. `kv_router_args` does not
import `router_args`, so composing `WorkerRouterConfig` there introduces no
cycle.

Call sites state their intent verbatim: the frontend keeps `"round-robin"` and
the frontend-only arguments; `counter_worker` keeps both, preserving its prior
behavior exactly; only the worker helper passes None/False.
